// src/app/pages/tipotrabajos/components/tipotrabajos-table/tipotrabajos.interface.ts
export interface TipoTrabajosInterface {
    claveauth?: string;
    nicknameauth?: string;
    usuarioauth?: string;
    idtipotrabajo?: number;
    descripcion: string;
    direccion: string;
    medidasterreno: string;
    medidasconstruccion: string;
    fechainicio: string;
    fechafin: string;
    idtipotipotrabajo: number;
    presupuesto: number;
    idrazonsocialcliente: number;
    idrazonsocialcontratista: number;
    idrazonsocialconstructor: number;
    idrazonsocialasociado: number;
    posiciongps: string;
    idestatustipotrabajo: number;
    observaciones: string;
    desctipotipotrabajo?: string;
    direccioncliente?: string;
    razonsocialcontratista?: string;
    razonsocialconstructor?: string;
    razonsocialasociado?: string;
    claveestatustipotrabajo?: string;
    tipotipotrabajo?: string;
}